Moved Filesystem Properties and DiscoverableDirectories into HTTPBuildableDirectoryMap namespace.
